Use !r for unknown values in exceptions
When raising an exception due to a wrong value being passed as an event
type, we do not know what the type of the value is—it should be a
string, but given that it is already known to be wrong, it might not be.
Use !r in the exception message so the application developer can better
see the unexpected value.