Add mutexes to send and receive
According to
<https://github.com/django/asgiref/issues/436#issuecomment-1897968476>,
it is legal for an application to issue multiple concurrent calls to the
`send` or `receive` callable. In order to avoid corruption at the SCGI
protocol layer due to this, and given that `asyncio` doesn’t appear to
make any documented guarantees of the form “the task that receives the
first data from a socket will also be the first task woken up”, a mutex
seems to be essential here.